requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
requests==2.32.3
python-dotenv==1.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-huggingface==0.0.3
langchain-openai==0.1.17
chromadb==0.4.24
pypdf==4.2.0
httpx==0.27.0